feat(Video): video domain model, videoTable [OMNI-55]

+++ backend/src/main/kotlin/com/v39a/omni/feature/video/domain/Video.kt
@@ -1,13 +1,12 @@
 package com.v39a.omni.feature.video.domain
 
 import kotlinx.datetime.LocalDateTime
-import java.math.BigInteger
 import java.util.UUID
 
 
 data class Video(val id: UUID, val path: String, val status: VideoStatus?,
                  val title: String,
-                 val durationSeconds: BigInteger,
+                 val durationSeconds: Int,
                  val thumbnailPath: String,
                  val createdAt: LocalDateTime,
 )
